pyproject.toml
[tool.poetry]
name = "qt"
version = "1.0.0"
description = "fast_api demo on the fly"
authors = ["vortexdude <nitinnamdeo456@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.105.0"
uvicorn = "^0.25.0"
motor = "2.5.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.6"
bcrypt = {version = "4.0.1", extras = ["bcrypt"]}
passlib = {extras = ["bcrpt"], version = "^1.7.4"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
